+++ src/main/java/spring/springserver/domain/community/comment/data/request/CreateCommentRequest.kt
@@ -1,0 +1,10 @@
+package spring.springserver.domain.community.comment.data.request
+
+import jakarta.validation.constraints.NotBlank
+
+data class CreateCommentRequest(
+    val postId: Long,
+
+    @field:NotBlank
+    val content: String,
+)
