--- src/main/java/spring/springserver/domain/community/data/request/CommunityCommentLikeRequest.kt
@@ -1,8 +1,0 @@
-package spring.springserver.domain.community.data.request
-
-import jakarta.validation.constraints.NotNull
-
-data class CommunityCommentLikeRequest(
-    @field:NotNull
-    val commentId: Long,
-)
