--- src/main/java/spring/springserver/domain/community/data/request/CreateCommentRequest.kt
@@ -1,10 +1,0 @@
-package spring.springserver.domain.community.data.request
-
-import jakarta.validation.constraints.NotBlank
-
-data class CreateCommentRequest(
-    val postId: Long,
-
-    @field:NotBlank
-    val content: String,
-)
