--- src/main/java/spring/springserver/domain/community/data/request/UpdateCommentRequest.kt
@@ -1,10 +1,0 @@
-package spring.springserver.domain.community.data.request
-
-import jakarta.validation.constraints.NotBlank
-
-data class UpdateCommentRequest(
-    val commentId: Long,
-
-    @field:NotBlank
-    val content: String,
-)
