--- src/main/java/spring/springserver/domain/community/data/response/DeleteResponse.kt
@@ -1,13 +1,0 @@
-package spring.springserver.domain.community.data.response
-
-data class DeleteResponse(
-    val message: String
-) {
-    companion object {
-
-        fun ok(): DeleteResponse {
-
-            return DeleteResponse("삭제되었습니다.")
-        }
-    }
-}
