+++ src/main/java/spring/springserver/domain/community/like/data/request/CommunityCommentLikeRequest.kt
@@ -1,0 +1,8 @@
+package spring.springserver.domain.community.like.data.request
+
+import jakarta.validation.constraints.NotNull
+
+data class CommunityCommentLikeRequest(
+    @field:NotNull
+    val commentId: Long,
+)
