+++ src/main/java/spring/springserver/domain/community/like/data/response/CommunityLikeResponse.kt
@@ -1,0 +1,21 @@
+package spring.springserver.domain.community.like.data.response
+
+data class CommunityLikeResponse(
+    val targetId: Long,
+
+    val likeCount: Long,
+
+    val message: String,
+) {
+    companion object {
+
+        fun of(targetId: Long, likeCount: Long, message: String): CommunityLikeResponse {
+
+            return CommunityLikeResponse(
+                targetId,
+                likeCount,
+                message,
+            )
+        }
+    }
+}
