+++ src/main/java/spring/springserver/domain/community/post/data/request/UpdatePostRequest.kt
@@ -1,0 +1,14 @@
+package spring.springserver.domain.community.post.data.request
+
+import jakarta.validation.constraints.NotBlank
+
+data class UpdatePostRequest(
+    val postId: Long,
+
+    @field:NotBlank
+    val title: String,
+
+    val content: String?,
+
+    val fileUrl: String?,
+)
