--- src/main/java/spring/springserver/domain/community/repository/CommunityCommentRepository.kt
@@ -1,18 +1,0 @@
-package spring.springserver.domain.community.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import spring.springserver.domain.community.entity.CommunityComment
-import java.time.LocalDateTime
-
-interface CommunityCommentRepository : JpaRepository<CommunityComment, Long> {
-
-    fun findByIdAndDeletedAtIsNull(id: Long): CommunityComment?
-
-    fun findAllByCommunityPostIdAndDeletedAtIsNullOrderByCreatedAtAsc(communityPostId: Long): List<CommunityComment>
-
-    fun countByCommunityPostIdAndDeletedAtIsNull(communityPostId: Long): Long
-
-    fun findAllByDeletedAtBefore(deletedAt: LocalDateTime): List<CommunityComment>
-
-    fun findAllByCommunityPostIdIn(communityPostIds: Collection<Long>): List<CommunityComment>
-}
